+++ llama_stack/distribution/configure_container.sh
