+++ llama_stack/distribution/datatypes.py
@@ -168,7 +168,7 @@ def remote_provider_spec(
     config_class = (
         adapter.config_class
         if adapter and adapter.config_class
-        else "llama_stack.core.datatypes.RemoteProviderConfig"
+        else "llama_stack.distribution.datatypes.RemoteProviderConfig"
     )
     provider_id = remote_provider_id(adapter.adapter_id) if adapter else "remote"
