+++ llama_stack/distribution/server/__init__.py
@@ -1,0 +1,12 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the terms described in the LICENSE file in
+# the root directory of this source tree.
+
+import fire
+
+from .server import main
+
+if __name__ == __main__:
+    fire.Fire(main)
