+++ llama_stack/providers/adapters/memory/chroma/__init__.py
@@ -4,7 +4,7 @@
 # This source code is licensed under the terms described in the LICENSE file in
 # the root directory of this source tree.
 
-from llama_stack.core.datatypes import RemoteProviderConfig
+from llama_stack.distribution.datatypes import RemoteProviderConfig
 
 
 async def get_adapter_impl(config: RemoteProviderConfig, _deps):
